Add the model viewer, and fix index buffers that were never filled
A Models/Zones switch in the top-left replaces the whole window.
ModelBrowser owns its own sidebar and viewport and shares only the FFXI
install path with the zone side. 2,473 NPC, monster and object models,
searchable and filterable by category.

DatFile.ts is the loader Vanalytics had and the original port left
behind. Everything it calls -- parseVertexBlock, parseSkeleton,
parseTextureBlock -- was already in the repo and unused. Animation
blocks are detected but not decoded, so models sit in bind pose.

MeshParser allocated its index buffers at the right length and never
filled them, so every index was 0 and every triangle degenerate. Models
rendered nothing while looking entirely healthy: meshes in the scene,
camera aimed at them, and renderer.info reporting 2,358 triangles --
that counter counts submitted geometry, not surviving fragments.
Vanalytics never hit it because it draws these meshes non-indexed.

Also fixes a StrictMode trap: creating three.js resources in useMemo and
disposing them in a useEffect cleanup means React's simulated unmount
disposes everything before the real mount renders it.

Verified on Goblin 1 (2,358 tris) and Mandragora (988 tris) by measuring
drawn pixels, not by eye.

+++ src/renderer/src/lib/ffxi-dat/types.ts
@@ -4,7 +4,7 @@ export interface ParsedMesh {
   vertices: Float32Array
   normals: Float32Array
   uvs: Float32Array
-  indices: Uint16Array
+  indices: Uint32Array
   boneIndices: Uint8Array
   boneWeights: Float32Array
   materialIndex: number
